+++ bobtemplates/plone/behavior.py
@@ -1,13 +1,11 @@
 # -*- coding: utf-8 -*-
-from bobtemplates.plone.base import base_prepare_renderer
-from bobtemplates.plone.base import echo
-from bobtemplates.plone.base import git_commit
-from bobtemplates.plone.base import update_file
-from bobtemplates.plone.utils import run_isort
-from lxml import etree
+import os
 
 import case_conversion as cc
-import os
+from lxml import etree
+
+from bobtemplates.plone.base import base_prepare_renderer, echo, git_commit, update_file
+from bobtemplates.plone.utils import run_black, run_isort
 
 
 def _update_package_configure_zcml(configurator):
@@ -99,6 +97,7 @@ def post_renderer(configurator):
     _update_package_configure_zcml(configurator)
     _update_behaviors_configure_zcml(configurator)
     run_isort(configurator)
+    run_black(configurator)
     git_commit(
         configurator,
         "Add behavior: {0}".format(