+++ bobtemplates/plone/svelte_app.py
@@ -1,15 +1,13 @@
 # -*- coding: utf-8 -*-
 
-from bobtemplates.plone.base import base_prepare_renderer
-from bobtemplates.plone.base import echo
-from bobtemplates.plone.base import git_commit
-from bobtemplates.plone.base import update_file
-from bobtemplates.plone.utils import run_isort
+import re
+
+import case_conversion as cc
 from lxml import etree
 from mrbob.bobexceptions import ValidationError
 
-import case_conversion as cc
-import re
+from bobtemplates.plone.base import base_prepare_renderer, echo, git_commit, update_file
+from bobtemplates.plone.utils import run_black, run_isort
 
 
 def check_name(configurator, question, answer):
@@ -69,6 +67,7 @@ def post_renderer(configurator):
     """Post rendering."""
     _update_configure_zcml(configurator)
     run_isort(configurator)
+    run_black(configurator)
     git_commit(
         configurator,
         "Add Svelte app: in svelte_apps/{0}".format(