+++ bobtemplates/plone/view.py
@@ -1,19 +1,21 @@
 # -*- coding: utf-8 -*-
 """Generate view."""
 
-from bobtemplates.plone.base import base_prepare_renderer
-from bobtemplates.plone.base import echo
-from bobtemplates.plone.base import git_commit
-from bobtemplates.plone.base import update_file
-from bobtemplates.plone.base import ZCML_NAMESPACES
-from bobtemplates.plone.utils import run_isort
-from lxml import etree
-from mrbob.bobexceptions import SkipQuestion
-from mrbob.bobexceptions import ValidationError
+import os
 
 import case_conversion as cc
-import os
 import six
+from lxml import etree
+from mrbob.bobexceptions import SkipQuestion, ValidationError
+
+from bobtemplates.plone.base import (
+    ZCML_NAMESPACES,
+    base_prepare_renderer,
+    echo,
+    git_commit,
+    update_file,
+)
+from bobtemplates.plone.utils import run_black, run_isort
 
 
 def get_view_name_from_python_class(configurator, question):
@@ -231,6 +233,7 @@ def post_renderer(configurator):
     _update_views_configure_zcml(configurator)
     _delete_unwanted_files(configurator)
     run_isort(configurator)
+    run_black(configurator)
     git_commit(
         configurator,
         "Add view: {0}".format(