+++ bobtemplates/plone/viewlet.py
@@ -1,16 +1,19 @@
 # -*- coding: utf-8 -*-
 """Generate view."""
 
-from bobtemplates.plone.base import base_prepare_renderer
-from bobtemplates.plone.base import git_commit
-from bobtemplates.plone.base import update_file
-from bobtemplates.plone.base import ZCML_NAMESPACES
-from bobtemplates.plone.utils import run_isort
+import os
+
+import case_conversion as cc
 from lxml import etree
 from mrbob.bobexceptions import SkipQuestion
 
-import case_conversion as cc
-import os
+from bobtemplates.plone.base import (
+    ZCML_NAMESPACES,
+    base_prepare_renderer,
+    git_commit,
+    update_file,
+)
+from bobtemplates.plone.utils import run_black, run_isort
 
 
 def get_view_name_from_python_class(configurator, question):
@@ -179,6 +182,7 @@ def post_renderer(configurator):
     _update_viewlets_configure_zcml(configurator)
     _delete_unwanted_files(configurator)
     run_isort(configurator)
+    run_black(configurator)
     git_commit(
         configurator,
         "Add viewlet: {0}".format(