+++ skeleton-tests/conftest.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import os
+
 import pytest
 
 
@@ -14,13 +15,13 @@ class Config(object):
         self.subtemplates = []
 
 
-@pytest.fixture(scope='module')
+@pytest.fixture(scope="module")
 def config():
-    skeleton_tox_env_parts = os.environ.get('ENVNAME').split('-')
+    skeleton_tox_env_parts = os.environ.get("ENVNAME").split("-")
     config = Config()
-    config.verbose = bool(os.environ.get('VERBOSE'))
-    config.version = os.environ.get('VERSION', '5.2.1')
-    config.skeleton_tox_env = u'{0}-{1},{0}-lint'.format(
+    config.verbose = bool(os.environ.get("VERBOSE"))
+    config.version = os.environ.get("VERSION", "5.2.1")
+    config.skeleton_tox_env = u"{0}-{1},{0}-lint".format(
         skeleton_tox_env_parts[0],
         skeleton_tox_env_parts[2],
     )